requirements.txt
# AWS CDK and core dependencies
aws-cdk-lib==2.160.0
constructs>=10.0.0,<11.0.0

# Development and testing
pytest==8.3.3
pytest-cov==5.0.0
pytest-mock==3.14.0
moto==5.0.18
black==24.10.0
flake8==7.1.1
mypy==1.13.0
boto3-stubs[dynamodb,s3,secretsmanager]==1.35.36
